Moved and update Users and Role Entities
