+++ src/main/java/com/oblitus/serviceApp/Security/Entities/User.java
@@ -1,8 +1,8 @@
-package com.oblitus.serviceApp.Security;
+package com.oblitus.serviceApp.Security.Entities;
 
 import com.oblitus.serviceApp.Abstracts.EntityBase;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
@@ -16,6 +16,7 @@ import java.util.Collection;
 @Entity
 @Table(name = "Users")
 public class User extends EntityBase {
+    @Column(nullable = false)
     public String Email;
     @OneToMany(targetEntity = Role.class)
     public Collection<Role> Roles;
